docs(fix-phase-01): phase report + Verifier hardening — PHASE_COMPLETE
Phase-end: compiled fix-phase-01-report.md (Verifier YES, 88% coverage, 17/17
E2E, Docker 6/6, no Critical/High). Addressed both Low Verifier findings:
- Order environments by (int)Name (robust if ordering ever moves to SQL).
- Tighten cross-org environments test to assert exactly 404.

159 backend tests pass (125 unit + 34 integration).

+++ backend/Flowamaz.Application/Workspace/Services/WorkspaceService.cs
@@ -146,7 +146,10 @@ public sealed class WorkspaceService : IWorkspaceService
         try
         {
             var environments = await _workspaceRepository.GetEnvironmentsAsync(workspaceId, cancellationToken);
-            var ordered = environments.OrderBy(e => e.Name).ToList();
+            // Order by the enum's numeric value (Dev=0, Staging=1, Production=2). The Name column is
+            // persisted as a string, so ordering must stay client-side — a SQL OrderBy would sort
+            // alphabetically (Dev, Production, Staging) and silently break the contract.
+            var ordered = environments.OrderBy(e => (int)e.Name).ToList();
             _logger.LogDebug(
                 "WorkspaceService.GetEnvironmentsAsync exit workspaceId={WorkspaceId} count={Count}", workspaceId, ordered.Count);
             return ordered;
